refactor: extracted UndecidedValue into undecided_value.py
