+++ src/inline_snapshot/_snapshot/min_max_value.py
@@ -2,12 +2,12 @@ from typing import Iterator
 
 from .._change import Change
 from .._change import Replace
-from .._inline_snapshot import clone
-from .._inline_snapshot import GenericValue
-from .._inline_snapshot import ignore_old_value
 from .._sentinels import undefined
 from .._utils import value_to_token
 from ..global_state import state
+from .generic_value import clone
+from .generic_value import GenericValue
+from .generic_value import ignore_old_value
 
 
 class MinMaxValue(GenericValue):
